#30 Added Gradle + small refactoring
