+++ src/main/java/CelsiusFahrenheit.java
@@ -1,0 +1,14 @@
+
+public class CelsiusFahrenheit {
+    double celsiusTemp;
+
+    public CelsiusFahrenheit (double temp) {
+        this.celsiusTemp = temp;
+
+    }
+
+    public double converter () {
+        return celsiusTemp * 1.8 + 32;
+    }
+
+}
